refactor: 추상클래스 모듈화
