+++ app/Constants/SocialConstant.php
@@ -5,4 +5,5 @@ namespace App\Constants;
 class SocialConstant
 {
     public const GOOGLE = "Google";
+    public const NAVER  = "Naver";
 }
